Dockerfiles/spark-master/requirements.txt
absl-py==0.13.0
astunparse==1.6.3
beautifulsoup4==4.6.0
bidict==0.21.2
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cycler==0.10.0
dnspython==1.16.0
h11==0.12.0
h2==3.2.0
h5py==3.1.0
hangul-utils==0.2
hpack==3.0.0
httpcore==0.12.3
httpx==0.16.1
hyperframe==5.2.0
idna==2.10
itsdangerous==2.0.1
jamo==0.4.1
jpype1==0.7.5
Jinja2==3.0.1
kafka-python==2.0.2
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
konlpy==0.5.2
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib
numpy==1.19.3
oauthlib==3.1.1
pafy==0.5.5
pandas
Pillow==8.2.0
pycparser==2.20
pyparsing==2.4.7
PySocks==1.7.1
pytchat==0.5.3
pytz==2021.1
python-dateutil==2.8.1
python-engineio==3.13.2
python-socketio==4.6.0
requests==2.25.1
requests-oauthlib==1.3.0
rfc3986==1.5.0
six==1.15.0
sniffio==1.2.0
tensorflow-gpu==2.5.0
tweepy==3.10.0
urllib3==1.26.5
websockets==9.1
Werkzeug==2.0.1
zope.event==4.5.0
zope.interface==5.4.0
